Align controller state setter with persistence hook
